+++ src/main/java/com/ureca/snac/auth/dto/TokenDto.java
@@ -1,0 +1,11 @@
+package com.ureca.snac.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenDto {
+    private String accessToken;
+    private String refreshToken;
+}
