+++ src/main/java/com/ureca/snac/auth/service/ReissueService.java
@@ -1,8 +1,10 @@
 package com.ureca.snac.auth.service;
 
+import com.ureca.snac.auth.dto.TokenDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface ReissueService {
-    void reissue(HttpServletRequest request, HttpServletResponse response);
+
+    TokenDto reissue(String refresh);
 }
