+++ src/main/java/com/ureca/snac/auth/service/ReissueServiceImpl.java
@@ -1,16 +1,13 @@
 package com.ureca.snac.auth.service;
 
+import com.ureca.snac.auth.dto.TokenDto;
 import com.ureca.snac.auth.jwt.JWTUtil;
 import com.ureca.snac.auth.refresh.Refresh;
 import com.ureca.snac.auth.repository.RefreshRepository;
 import com.ureca.snac.common.BaseCode;
 import com.ureca.snac.common.exception.BusinessException;
 import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -21,18 +18,9 @@ public class ReissueServiceImpl implements ReissueService {
     private final RefreshRepository refreshRepository;
 
     @Override
-    public void reissue(HttpServletRequest request, HttpServletResponse response) {
-        // 1. 리프레시 토큰 쿠키로부터 가져오고
-        String refresh = null;
-        Cookie[] cookies = request.getCookies();
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                if ("refresh".equals(cookie.getName())) {
-                    refresh = cookie.getValue();
-                }
-            }
-        }
+    public TokenDto reissue(String refresh) {
 
+        // 1. 리프레시 토큰 있는지 검증
         if (refresh == null) {
             throw new BusinessException(BaseCode.REFRESH_TOKEN_NULL);
         }
@@ -41,9 +29,7 @@ public class ReissueServiceImpl implements ReissueService {
         try {
             jwtUtil.isExpired(refresh);
         } catch (ExpiredJwtException e) {
-
             refreshRepository.findByRefresh(refresh).ifPresent(refreshRepository::delete);
-
             throw new BusinessException(BaseCode.REFRESH_TOKEN_EXPIRED);
         }
 
@@ -68,19 +54,9 @@ public class ReissueServiceImpl implements ReissueService {
 
 
         // 5. 기존 리프레시 토큰 레디스에서 삭제, 새 거 저장 => 생각해보니까 굳이 삭제 할 필요가 없고 덮어씌우면 되어서 코드 변경
-
         refreshRepository.save(new Refresh(username, newRefresh));
 
-        response.setHeader(HttpHeaders.AUTHORIZATION, "Bearer " + newAccess);
-        response.addCookie(createCookie("refresh", newRefresh));
+        return new TokenDto(newAccess, newRefresh);
     }
 
-    private Cookie createCookie(String key, String value) {
-        Cookie cookie = new Cookie(key, value);
-        cookie.setMaxAge(24 * 60 * 60);
-        cookie.setPath("/api");
-        cookie.setHttpOnly(true);
-        cookie.setDomain("localhost");
-        return cookie;
-    }
 }
